requirements.txt
selenium
python-dotenv
rasa